10/10 Cold Meter update
updated cold meter functionality
- cold meter runs down when away from heat source
- when cold reaches zero, player enters Freezing state, health depletes until death
- when in Freezing state, movement speed is halved

Added warm sources,
-bonfires will replenish cold meter (should rename to 'Warmth Meter')

Added a simple reticle to help players track where their rock throws will go
- this is no way entirely accurate, but does assist...

+++ Assets/Scripts/HealthBar.cs
@@ -9,13 +9,13 @@ namespace MBF
     {
         public Slider slider;
 
-        public void SetMaxHealth(int maxhealth)
+        public void SetMaxHealth(float maxhealth)
         {
             slider.maxValue = maxhealth;
             slider.value = maxhealth;
         }
 
-        public void SetCurrenthealth(int currentHealth)
+        public void SetCurrenthealth(float currentHealth)
         {
             slider.value = currentHealth;
         }
